requirement.txt
flask
flask-classy
Flask-Testing
Pillow
xxhash
six
